perf(mrmr): compact int8/int16 storage for the categorized codes matrix
The categorized factors_data is the big (n_samples, n_features) working matrix -- at scale (e.g. 795k rows x 496
features) it is ~1.58 GB as int32. But it only ever holds per-column BIN INDICES (0..nbins-1 + a NaN bin), never JOINT
class ids, so int32 is 4x wider than needed. After categorization (range-checked, one min/max pass) the matrix is downcast
to int8 (codes fit -128..127, the common nbins<=~127 case -> 4x) or int16 (a high-cardinality categorical -> 2x).

Selection-EQUIVALENT: the code VALUES are unchanged, and every consumer casts the storage UP for JOINT math -- merge_vars
does factors_data[:, j].astype(dtype) and the GPU path .astype(np.int32) -- so deep joints (nbins^order) never overflow
the narrow storage. Engineered-code appends in the FE steps (_step_score, the hinge append) cast down to data.dtype so the
matrix stays compact through the greedy loop. Default on; opt out with MLFRAME_MRMR_COMPACT_CODES=0.

Tests: selection identical compact ON vs OFF (fe_max_steps 0 and 1, exercising the FE append path); int32->int8 downcast
preserves values + is 4x smaller. Group-aware MI + CMI-cuda wiring regression (20 passed) unaffected.

+++ src/mlframe/feature_selection/filters/_mrmr_fe_step/_step_score.py
@@ -870,7 +870,9 @@ def materialise_and_finalise_fe_candidates(
                             dtype=self.quantization_dtype,
                         )
                     _n_cols_before_esc = len(cols)
-                    data = np.append(data, _esc_new_codes, axis=1)
+                    # Preserve the compact codes dtype (COMPACT CODES STORAGE in _fit_impl_core): nbins-bounded codes,
+                    # cast down to data.dtype rather than upcasting the whole matrix back to int32.
+                    data = np.append(data, _esc_new_codes.astype(data.dtype, copy=False), axis=1)
                     nbins = np.concatenate([
                         np.asarray(nbins),
                         np.asarray([self.quantization_nbins] * len(_esc_admitted), dtype=np.asarray(nbins).dtype),
@@ -975,7 +977,7 @@ def materialise_and_finalise_fe_candidates(
                         dtype=self.quantization_dtype,
                     )
                 _n_cols_before_fz = len(cols)
-                data = np.append(data, _fz_codes, axis=1)
+                data = np.append(data, _fz_codes.astype(data.dtype, copy=False), axis=1)
                 nbins = np.concatenate([
                     np.asarray(nbins),
                     np.asarray([self.quantization_nbins] * len(_fused), dtype=np.asarray(nbins).dtype),
